Consolidate code into monolithic module
