--- arianna_c/model.py
@@ -1,131 +1,0 @@
-import math
-from dataclasses import dataclass
-
-import torch
-import torch.nn as nn
-
-from .tokenizer import tokenizer
-
-
-@dataclass
-class AriannaCConfig:
-    """Configuration for the Arianna-C transformer."""
-
-    block_size: int = 1024
-    vocab_size: int | None = None
-    n_layer: int = 12
-    n_head: int = 12
-    n_embd: int = 768
-    dropout: float = 0.0
-
-    def __post_init__(self) -> None:
-        if self.vocab_size is None:
-            self.vocab_size = tokenizer.vocab_size
-
-
-class CausalSelfAttention(nn.Module):
-    def __init__(self, config: AriannaCConfig):
-        super().__init__()
-        assert config.n_embd % config.n_head == 0
-        self.key = nn.Linear(config.n_embd, config.n_embd, bias=False)
-        self.query = nn.Linear(config.n_embd, config.n_embd, bias=False)
-        self.value = nn.Linear(config.n_embd, config.n_embd, bias=False)
-        self.proj = nn.Linear(config.n_embd, config.n_embd)
-        self.attn_dropout = nn.Dropout(config.dropout)
-        self.resid_dropout = nn.Dropout(config.dropout)
-        self.n_head = config.n_head
-        self.head_dim = config.n_embd // config.n_head
-        self.register_buffer(
-            "bias",
-            torch.tril(torch.ones(config.block_size, config.block_size)).view(
-                1, 1, config.block_size, config.block_size
-            ),
-            persistent=False,
-        )
-
-    def forward(self, x: torch.Tensor) -> torch.Tensor:
-        B, T, C = x.size()
-        k = self.key(x).view(B, T, self.n_head, self.head_dim).transpose(1, 2)
-        q = self.query(x).view(B, T, self.n_head, self.head_dim).transpose(1, 2)
-        v = self.value(x).view(B, T, self.n_head, self.head_dim).transpose(1, 2)
-
-        att = (q @ k.transpose(-2, -1)) * (1.0 / math.sqrt(self.head_dim))
-        att = att.masked_fill(self.bias[:, :, :T, :T] == 0, float("-inf"))
-        att = torch.softmax(att, dim=-1)
-        att = self.attn_dropout(att)
-
-        y = att @ v
-        y = y.transpose(1, 2).contiguous().view(B, T, C)
-        y = self.resid_dropout(self.proj(y))
-        return y
-
-
-class MLP(nn.Module):
-    def __init__(self, config: AriannaCConfig):
-        super().__init__()
-        self.fc = nn.Linear(config.n_embd, 4 * config.n_embd)
-        self.proj = nn.Linear(4 * config.n_embd, config.n_embd)
-        self.dropout = nn.Dropout(config.dropout)
-
-    def forward(self, x: torch.Tensor) -> torch.Tensor:
-        x = torch.nn.functional.gelu(self.fc(x))
-        x = self.proj(x)
-        return self.dropout(x)
-
-
-class Block(nn.Module):
-    def __init__(self, config: AriannaCConfig):
-        super().__init__()
-        self.ln1 = nn.LayerNorm(config.n_embd)
-        self.attn = CausalSelfAttention(config)
-        self.ln2 = nn.LayerNorm(config.n_embd)
-        self.mlp = MLP(config)
-
-    def forward(self, x: torch.Tensor) -> torch.Tensor:
-        x = x + self.attn(self.ln1(x))
-        x = x + self.mlp(self.ln2(x))
-        return x
-
-
-class AriannaC(nn.Module):
-    """A minimal GPT-style model inspired by nanoGPT."""
-
-    def __init__(self, config: AriannaCConfig):
-        super().__init__()
-        self.config = config
-        self.token_embedding = nn.Embedding(config.vocab_size, config.n_embd)
-        self.position_embedding = nn.Embedding(config.block_size, config.n_embd)
-        self.drop = nn.Dropout(config.dropout)
-        self.blocks = nn.ModuleList([Block(config) for _ in range(config.n_layer)])
-        self.ln_f = nn.LayerNorm(config.n_embd)
-        self.head = nn.Linear(config.n_embd, config.vocab_size, bias=False)
-        self.block_size = config.block_size
-
-    def forward(self, idx: torch.Tensor, targets: torch.Tensor | None = None):
-        B, T = idx.size()
-        if T > self.block_size:
-            raise ValueError("Cannot forward, sequence too long")
-        pos = torch.arange(0, T, dtype=torch.long, device=idx.device).unsqueeze(0)
-        tok_emb = self.token_embedding(idx)
-        pos_emb = self.position_embedding(pos)
-        x = self.drop(tok_emb + pos_emb)
-        for block in self.blocks:
-            x = block(x)
-        x = self.ln_f(x)
-        logits = self.head(x)
-        loss = None
-        if targets is not None:
-            loss = torch.nn.functional.cross_entropy(
-                logits.view(-1, logits.size(-1)), targets.view(-1)
-            )
-        return logits, loss
-
-    @torch.no_grad()
-    def generate(self, idx: torch.Tensor, max_new_tokens: int) -> torch.Tensor:
-        for _ in range(max_new_tokens):
-            logits, _ = self(idx[:, -self.block_size :])
-            logits = logits[:, -1, :]
-            probs = torch.softmax(logits, dim=-1)
-            idx_next = torch.multinomial(probs, num_samples=1)
-            idx = torch.cat((idx, idx_next), dim=1)
-        return idx
