+++ setup.cfg
@@ -1,10 +1,8 @@
 [isort]
 profile = black
-known_first_party = arianna_c
+known_first_party = arianna_chain
 line_length = 119
 
 [flake8]
 max-line-length = 119
 extend-ignore = E203
-per-file-ignores =
-    __init__.py: F401
